chore: Move user classes down a bit
